[CPU] Add explicit float cast for CIs

+++ src/plugins/intel_cpu/src/nodes/multinomial.cpp
@@ -213,7 +213,7 @@ void Multinomial::execute_convert_type() {
         gen.seed(m_op_seed + gen());
     }
 
-    const auto gen_max = gen.max();
+    const auto gen_max = static_cast<float>(gen.max());
     std::generate(m_random_samples.begin(), m_random_samples.end(), [&]() {
         return static_cast<P>(static_cast<float>(gen())/gen_max);
     });
